+++ js/util.js
@@ -1,3 +1,8 @@
+const getRandomInteger = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export { getRandomInteger, getRandomArrayElement };
+
 const checkStringLength = (string, length) => (string.length) <= length;
 
 checkStringLength();
